add the fact that bombs can remove tiles, but still I have to udpate the Viewer for the tiles with a StageViewer

+++ src/BombModel.java
@@ -14,7 +14,7 @@ public class BombModel extends EntityModel{
         time = timer;
     }
 
-    public double getBlastRadius() {
+    public int getBlastRadius() {
         return blastRadius;
     }
